requirements.txt
markdown==3.6
numpy==2.0.1
psycopg2-binary==2.9.9 